LPS-32723 Move logic for checking membership when login to a post processor so it can be unplugged in case we don't want to use membership policies
