+++ portal-impl/src/portal.properties
@@ -4142,7 +4142,7 @@
     # Login event
     #
     login.events.pre=com.liferay.portal.events.LoginPreAction
-    login.events.post=com.liferay.portal.events.ChannelLoginPostAction,com.liferay.portal.events.DefaultLandingPageAction,com.liferay.portal.events.LoginPostAction
+    login.events.post=com.liferay.portal.events.ChannelLoginPostAction,com.liferay.portal.events.DefaultLandingPageAction,com.liferay.portal.events.MembershipPolicyLoginPostAction,com.liferay.portal.events.LoginPostAction
 
     #
     # Logout event
